CVTO/requirements.txt
# CTVO Pipeline Requirements
# Core dependencies for the hybrid virtual try-on pipeline

# PyTorch and related
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# PyTorch Lightning
pytorch-lightning>=1.8.0
lightning-bolts>=0.6.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0
scikit-image>=0.19.0

# Deep Learning Utilities
timm>=0.6.0
einops>=0.4.0
kornia>=0.6.0

# ONNX Runtime (for Stage 1 parsing models)
onnxruntime>=1.12.0
onnx>=1.12.0

# Data Processing
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0

# Configuration and Logging
PyYAML>=6.0
hydra-core>=1.2.0
wandb>=0.13.0
tensorboard>=2.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
pre-commit>=2.19.0

# Optional: CUDA support
# torch-audio>=0.12.0+cu116 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision>=0.13.0+cu116 -f https://download.pytorch.org/whl/torch_stable.html

# Optional: Advanced features
# accelerate>=0.12.0  # For distributed training
# deepspeed>=0.7.0     # For large model training
# flash-attn>=2.0.0    # For attention optimization
